refactor: upgrade React 19 RC
This commit updates the Vite configuration file (vite.config.ts) to include the babel-plugin-react-compiler plugin. This plugin is necessary for the new ReactCompilerConfig feature that has been added to the configuration. By including this plugin, the ReactCompilerConfig will be properly processed during the build process.

The changes include:
- Importing the react-compiler plugin from "@vitejs/plugin-react"
- Adding the react-compiler plugin to the plugins array in the defineConfig function
- Configuring the babel plugin with the ReactCompilerConfig

These changes are necessary to ensure that the ReactCompilerConfig is properly applied to the project.